Implementation of ScrollableStackView
This component uses a UIScrollView and a UIStackView under the hood, and
allows the caller to initialize it with a given `NSLayoutConstraint.Axis.

To add items to the underlying `stackView`, simply access it and use the
`addArrangedSubView` method. The `stackView` and `scrollView` are public
properties so the stackView or scrollView properties can be configured.